+++ src/application/index.ts
@@ -1,3 +1,4 @@
 export { GetCurrentPositionUseCase } from './use-cases/GetCurrentPositionUseCase';
 export { WatchPositionUseCase } from './use-cases/WatchPositionUseCase';
 export type { GetCurrentPositionOutput } from './dtos/GetCurrentPositionOutput';
+export { GeolocationService } from './services/GeolocationService';